API

Radii polynomial

RadiiPolynomial.roots_radii_polynomial — Function

roots_radii_polynomial(Y::Interval{T}, Z₁::Interval{T}, r₀::T) where {T<:Real}

Return the interval of existence garanted by the Radii Polynomial Theorem from the root of the polynomial Y + (Z₁-1)r.

roots_radii_polynomial(Y::Interval{T}, Z₁::Interval{T}, Z₂::Interval{T}, r₀::T) where {T<:Real}

Return the interval of existence garanted by the Radii Polynomial Theorem from the root(s) of the polynomial Y + (Z₁-1)r + Z₂r²/2.

Spaces

RadiiPolynomial.SequenceSpace — Type

UnivariateSpace <: SequenceSpace

Abstract type for sequence spaces.

RadiiPolynomial.UnivariateSpace — Type

UnivariateSpace <: SequenceSpace

Abstract type for univariate sequence spaces.

RadiiPolynomial.Taylor — Type

Taylor <: UnivariateSpace

Taylor sequence space whose elements are Taylor sequences of a prescribed order.

Fields:

order :: Int

RadiiPolynomial.Fourier — Type

Fourier{T} <: UnivariateSpace

Fourier sequence space whose elements are Fourier sequences of a prescribed order and frequency.

Fields:

order :: Int

frequency :: T

RadiiPolynomial.Chebyshev — Type

Chebyshev <: UnivariateSpace

Chebyshev sequence space whose elements are Chebyshev sequences of a prescribed order.

Fields:

order :: Int

RadiiPolynomial.TensorSpace — Type

TensorSpace{T<:NTuple{N,UnivariateSpace} where {N}} <: SequenceSpace

Multivariate SequenceSpace resulting from the tensor product of some UnivariateSpace.

Fields:

spaces :: T

Sequences

RadiiPolynomial.Sequence — Type

Sequence{T<:SequenceSpace,S<:AbstractVector}

Compactly supported sequence of the given space.

Fields:

space :: T

coefficients :: S

RadiiPolynomial.project — Method

project(a::Sequence, space::SequenceSpace)

Return a Sequence representing a in space.

Base.selectdim — Method

selectdim(a::Sequence{TensorSpace{T}}, dim::Int, i::Int) where {N,T<:NTuple{N,UnivariateSpace}}

Return a Sequence in the space a.space[1:dim-1] ⊗ a.space[dim+1:N] and whose coefficients are a view of all the data a.coefficients where the index for dimension dim equals i.

RadiiPolynomial.shift — Function

shift(a::Sequence{<:Fourier}, from, to)

Return the shifted Fourier sequence representing a shifted Fourier series centered at to and was originally centered at from.

RadiiPolynomial.banach_algebra_rounding! — Function

banach_algebra_rounding!(a::Sequence, ν::Vector)

Return a rounded Sequence according to the decay ν (c.f. Computing Discrete Convolutions with Verified Accuracy Via Banach Algebras and the FFT).

Functionals

RadiiPolynomial.Functional — Type

Functional{T<:SequenceSpace,S<:AbstractVector}

Compactly supported functional with effective domain.

Fields:

domain :: T

coefficients :: S

RadiiPolynomial.project — Method

project(A::Functional, domain::SequenceSpace)

Return a Functional representing A with effective domain.

Base.selectdim — Method

selectdim(a::Functional{TensorSpace{T}}, dim::Int, i::Int) where {N,T<:NTuple{N,UnivariateSpace}}

Return a Functional with effective domain A.domain[1:dim-1] ⊗ A.domain[dim+1:N] and whose coefficients are a view of all the data A.coefficients where the index for dimension dim equals i.

Operators

RadiiPolynomial.Operator — Type

Operator{T<:SequenceSpace,S<:SequenceSpace,R<:AbstractMatrix}

Compactly supported operator with effective domain and range.

Fields:

domain :: T

range :: S

coefficients :: R

RadiiPolynomial.project — Method

project(A::Operator, domain::SequenceSpace, range::SequenceSpace)

Return an Operator representing A with effective domain and range.