Functionals

A Functional is a structure representing a compactly supported element of the dual space of a sequence space. More precisely, a Functional has an effective domain. The effective domain of a functional is defined as a SequenceSpace with the property that the functional and its restriction to the effective domain coincide.

Defining Functional with an effective domain allows different subspaces of a given sequence space to interact.

RadiiPolynomial.Functional — Type

Functional{T<:SequenceSpace,S<:AbstractVector}

Compactly supported functional with effective domain.

Fields:

domain :: T

coefficients :: S

Action

The action of a Functional is performed by the left product of a Sequence with a Functional.

julia> using RadiiPolynomial

julia> A = Functional(Taylor(2), [1.0, 2.0, 3.0]);

julia> A*Sequence(Taylor(1), [1.0, 2.0])
5.0

julia> A*Sequence(Taylor(2), [1.0, 2.0, 3.0])
14.0

julia> A*Sequence(Taylor(3), [1.0, 2.0, 3.0, 4.0])
14.0

Note

Functional is callable: (A::Functional)(b::Sequence) = A*b.

Arithmetic

The arithmetic operations +,-,*,^ are implemented along with the convenient bar operations +̄,-̄,*̄,^̄ (+\bar<TAB>, -\bar<TAB>, *\bar<TAB>, ^\bar<TAB>).

julia> using RadiiPolynomial

julia> A = Functional(Taylor(2) ⊗ Chebyshev(1), rand(1:3, length(Taylor(2) ⊗ Chebyshev(1))));

julia> A.coefficients
6-element Array{Int64,1}:
 3
 2
 3
 3
 1
 2

julia> B = Functional(Taylor(1) ⊗ Chebyshev(2), rand(1:3, length(Taylor(1) ⊗ Chebyshev(2))));

julia> B.coefficients
6-element Array{Int64,1}:
 3
 1
 1
 3
 2
 2

julia> C = A + B;

julia> C.domain
Taylor(2) ⨂ Chebyshev(2)

julia> C.coefficients
9-element Array{Int64,1}:
 6
 3
 3
 4
 4
 2
 2
 2
 0

julia> C̄ = A +̄ B;

julia> C̄.domain
Taylor(1) ⨂ Chebyshev(1)

julia> C̄.coefficients
4-element Array{Int64,1}:
 6
 3
 4
 4