Spaces

The spaces under consideration are the ones where the elements are compactly supported sequences.

The abstract type SequenceSpace has the two subtypes UnivariateSpace and TensorSpace where the latter corresponds to the tensor products of some spaces in UnivariateSpace.

RadiiPolynomial.SequenceSpace — Type

UnivariateSpace <: SequenceSpace

Abstract type for sequence spaces.

RadiiPolynomial.UnivariateSpace — Type

UnivariateSpace <: SequenceSpace

Abstract type for univariate sequence spaces.

RadiiPolynomial.Taylor — Type

Taylor <: UnivariateSpace

Taylor sequence space whose elements are Taylor sequences of a prescribed order.

Fields:

order :: Int

RadiiPolynomial.Fourier — Type

Fourier{T} <: UnivariateSpace

Fourier sequence space whose elements are Fourier sequences of a prescribed order and frequency.

Fields:

order :: Int

frequency :: T

RadiiPolynomial.Chebyshev — Type

Chebyshev <: UnivariateSpace

Chebyshev sequence space whose elements are Chebyshev sequences of a prescribed order.

Fields:

order :: Int

RadiiPolynomial.TensorSpace — Type

TensorSpace{T<:NTuple{N,UnivariateSpace} where {N}} <: SequenceSpace

Multivariate SequenceSpace resulting from the tensor product of some UnivariateSpace.

Fields:

spaces :: T