API

RadiiPolynomial.roots_radii_polynomial — Function

roots_radii_polynomial(Y::Interval{T}, Z₁::Interval{T}, r₀::T) where {T<:Real}

Return the interval of existence garanted by the Radii Polynomial Theorem from the root of the polynomial Y + (Z₁-1)r.

roots_radii_polynomial(Y::Interval{T}, Z₁::Interval{T}, Z₂::Interval{T}, r₀::T) where {T<:Real}

Return the interval of existence garanted by the Radii Polynomial Theorem from the root(s) of the polynomial Y + (Z₁-1)r + Z₂r²/2.

RadiiPolynomial.FixedPointProblemFiniteDimension — Type

FixedPointProblemFiniteDimension{T₁,T₂,T₃,T₄,T₅,T₆}

Fields:

x₀ :: T₁

G :: T₂

DG :: T₃

D²G :: T₄

p :: T₅

r₀ :: T₆

RadiiPolynomial.ZeroFindingProblemFiniteDimension — Type

ZeroFindingProblemFiniteDimension{T₁,T₂,T₃,T₄,T₅,T₆,T₇}

Fields:

x₀ :: T₁

F :: T₂

DF :: T₃

A :: T₄

D²F :: T₅

p :: T₆

r₀ :: T₇

RadiiPolynomial.ZeroFindingProblemCategory1 — Type

ZeroFindingProblemCategory1{T₁,T₂,T₃,T₄,T₅,T₆,T₇,T₈,T₉,T₁₀}

Fields:

∞f :: T₁

Df :: T₂

D²_abs_f :: T₃

ᴺF :: T₄

ᴺDF :: T₅

ᴺAᴺ :: T₆

ᴺL∞_bound :: T₇

∞L⁻¹∞_bound :: T₈

ν :: T₉

r₀ :: T₁₀

RadiiPolynomial.TailProblem — Type

TailProblem{T₁,T₂,T₃,T₄,T₅,T₆}

Fields:

∞f :: T₁

Df :: T₂

D²_abs_f :: T₃

∞L∞_bound :: T₄

ν :: T₅

r₀ :: T₆

RadiiPolynomial.Y — Function

Y(pb::FixedPointProblemFiniteDimension)

Compute the norm $\| G(x_0) - x_0 \|_p$.

Y(pb::ZeroFindingProblemFiniteDimension)

Compute the norm $\| A F(x_0) \|_p$.

RadiiPolynomial.Z₁ — Function

Z₁(pb::FixedPointProblemFiniteDimension)

Compute the operator norm $\| D G(y) \|_p$ where $y = x_0$ or $y \in \overline{B_{r_0}(x_0)}$.

Z₁(pb::ZeroFindingProblemFiniteDimension)

Compute the operator norm $\| A D F(y) - I \|_p$ where $y = x_0$ or $y \in \overline{B_{r_0}(x_0)}$.

RadiiPolynomial.Z₂ — Function

Z₂(pb::FixedPointProblemFiniteDimension)

Compute the operator norm $\| D^2 G(y) \|_p$ where $y \in \overline{B_{r_0}(x_0)}$.

Z₂(pb::ZeroFindingProblemFiniteDimension)

Compute the operator norm $\| A D^2 F(y) \|_p$ where $y \in \overline{B_{r_0}(x_0)}$.

RadiiPolynomial.VectorSpace — Type

VectorSpace

Abstract type for all vector spaces.

RadiiPolynomial.ParameterSpace — Type

ParameterSpace <: VectorSpace

Space of parameters.

RadiiPolynomial.SequenceSpace — Type

SequenceSpace <: VectorSpace

Abstract type for sequence spaces.

RadiiPolynomial.UnivariateSpace — Type

UnivariateSpace <: SequenceSpace

Abstract type for univariate sequence spaces.

RadiiPolynomial.Taylor — Type

Taylor <: UnivariateSpace

Taylor sequence space whose elements are Taylor sequences of a prescribed order.

Fields:

order :: Int

RadiiPolynomial.Fourier — Type

Fourier{T} <: UnivariateSpace

Fourier sequence space whose elements are Fourier sequences of a prescribed order and frequency.

Fields:

order :: Int

frequency :: T

RadiiPolynomial.Chebyshev — Type

Chebyshev <: UnivariateSpace

Chebyshev sequence space whose elements are Chebyshev sequences of a prescribed order.

Fields:

order :: Int

RadiiPolynomial.TensorSpace — Type

TensorSpace{T<:NTuple{N,UnivariateSpace} where {N}} <: SequenceSpace

Multivariate SequenceSpace resulting from the tensor product of some UnivariateSpace.

Fields:

spaces :: T

Missing docstring.

Missing docstring for CartesianSpace. Check Documenter's build log for details.

RadiiPolynomial.Sequence — Type

Sequence{T<:VectorSpace,S<:AbstractVector}

Compactly supported sequence of the given space.

Fields:

space :: T

coefficients :: S

RadiiPolynomial.Operator — Type

Operator{T<:VectorSpace,S<:VectorSpace,R<:Union{AbstractMatrix,Factorization}}

Compactly supported operator with effective domain and codomain.

Fields:

domain :: T

codomain :: S

coefficients :: R