Norms

The choice of the Banach space to apply the Radii Polynomial Theorem (cf. Section Radii polynomial approach) is integral to the success of the computer-assisted proof. In practice, it is useful to tune the Banach space on the fly to adjust the norm estimates.

Accordingly, the spaces introduced in Section Vector spaces are not normed a priori. The norm of a Sequence or a LinearOperator is obtained via the functions norm and opnorm respectively; in both cases, one must specify a BanachSpace.

BanachSpace
├─ NormedCartesianSpace
├─ ℓ¹
├─ ℓ²
└─ ℓ∞

RadiiPolynomial.BanachSpace — Type

BanachSpace

Abstract type for all Banach spaces.

$\ell^1$, $\ell^2$ and $\ell^\infty$

Let $\mathscr{I}$ be a set of indices such that $\mathscr{I} \subset \mathbb{Z}^d$ for some $d \in \mathbb{N}$. Consider the weighted $\ell^1, \ell^2, \ell^\infty$ spaces defined by

\[\begin{aligned} \ell^1_w &:= \left\{ a \in \mathbb{C}^\mathscr{I} \, : \, | a |_{\ell^1_w} := \sum_{\alpha \in \mathscr{I}} |a_\alpha| w(\alpha) < +\infty \right\}, \\ \ell^2_w &:= \left\{ a \in \mathbb{C}^\mathscr{I} \, : \, | a |_{\ell^2_w} := \sqrt{\sum_{\alpha \in \mathscr{I}} |a_\alpha|^2 w(\alpha)} < +\infty \right\}, \\ \ell^\infty_w &:= \left\{ a \in \mathbb{C}^\mathscr{I} \, : \, | a |_{\ell^\infty_w} := \sup_{\alpha \in \mathscr{I}} | a_\alpha | w(\alpha) < +\infty \right\}, \end{aligned}\]

where $w : \mathscr{I} \to (0, +\infty)$ is a weight function.

The Banach spaces ℓ¹ (\ell<TAB>\^1<TAB>), ℓ² (\ell<TAB>\^2<TAB>) and ℓ∞ (\ell<TAB>\infty<TAB>) wraps one or multiple Weight.

Weight
├─ AlgebraicWeight
├─ BesselWeight
├─ GeometricWeight
└─ IdentityWeight

Given a set of indices $\mathscr{I}^\prime \subset \mathbb{Z}$:

an AlgebraicWeight of rate $s \geq 0$ is defined by $w(\alpha) := (1 + |\alpha|)^s$ for all $\alpha \in \mathscr{I}^\prime$.

a BesselWeight of rate $s \geq 0$ is defined by $w(\alpha) := (1 + |\alpha|)^s$ for all $\alpha \in \mathscr{I}^\prime$. This weight is specific to ℓ² and Fourier as it describes the Sobolev space $H^s$.

a GeometricWeight of rate $\nu > 0$ is defined by $w(\alpha) := \nu^{|\alpha|}$ for all $\alpha \in \mathscr{I}^\prime$.

an IdentityWeight is defined by $w(\alpha) := 1$ for all $\alpha \in \mathscr{I}^\prime$. This is the default weight for ℓ¹, ℓ² and ℓ∞.

julia> a = Sequence(Taylor(2), [1.0, 1.0, 1.0]); # 1 + x + x^2
julia> norm(a, ℓ¹(AlgebraicWeight(1.0)))6.0
julia> b = Sequence(Fourier(1, 1.0), [0.5, 0.0, 0.5]); # cos(x)
julia> norm(b, ℓ²(BesselWeight(2.0)))1.4142135623730951
julia> c = Sequence(Chebyshev(2), [1.0, 0.5, 0.5]); # 1 + 2(x/2 + (2x^2 - 1)/2)
julia> norm(c, ℓ∞()) # ℓ∞() == ℓ∞(IdentityWeight())1.0

In the context of a $d$-dimensional TensorSpace, one prescribes weights $w_1, \dots, w_d$ for each dimension. The weight is defined by $w(\alpha) = w_1(\alpha_1) \times \dots \times w_d(\alpha_d)$ for all $\alpha = (\alpha_1, \dots, \alpha_d) \in \mathscr{I}^{\prime\prime}$ where $\mathscr{I}^{\prime\prime} \subset \mathbb{Z}^d$ is the appropriate set of indices.

julia> a = Sequence(Taylor(2) ⊗ Fourier(2, 1.0) ⊗ Chebyshev(2), ones(3*5*3));
julia> norm(a, ℓ¹((AlgebraicWeight(1.0), GeometricWeight(2.0), IdentityWeight())))390.0

However, the $d$-dimensional version of BesselWeight is defined by $w(\alpha) := (1 + |\alpha_1| + \dots + |\alpha_d|)^s$ for all $\alpha = (\alpha_1, \dots, \alpha_d) \in \mathbb{Z}^d$. Only one BesselWeight is required for every Fourier space composing the TensorSpace.

julia> a = Sequence(Fourier(2, 1.0) ⊗ Fourier(3, 1.0), ones(5*7));
julia> norm(a, ℓ²(BesselWeight(2.0)))47.25462940284264

RadiiPolynomial.ℓ¹ — Type

ℓ¹{T<:Union{Weight,Tuple{Vararg{Weight}}}} <: BanachSpace

$\ell^1$ Banach space.

Fields:

weight :: T

RadiiPolynomial.ℓ² — Type

ℓ²{T<:Union{Weight,Tuple{Vararg{Weight}}}} <: BanachSpace

$\ell^2$ Banach space.

Fields:

weight :: T

RadiiPolynomial.ℓ∞ — Type

ℓ∞{T<:Union{Weight,Tuple{Vararg{Weight}}}} <: BanachSpace

$\ell^\infty$ Banach space.

Fields:

weight :: T

RadiiPolynomial.Weight — Type

Weight

Abstract type for all weights.

RadiiPolynomial.AlgebraicWeight — Type

AlgebraicWeight{T<:Real} <: Weight

Algebraic weight.

Fields:

rate :: T

RadiiPolynomial.BesselWeight — Type

BesselWeight{T<:Real} <: Weight

Bessel weight.

Fields:

rate :: T

RadiiPolynomial.GeometricWeight — Type

GeometricWeight{T<:Real} <: Weight

Geometric weight.

Fields:

rate :: T

RadiiPolynomial.IdentityWeight — Type

IdentityWeight <: Weight

Identity weight.

Normed cartesian space

For the norm of a CartesianSpace, one may use a NormedCartesianSpace to either:

use the same BanachSpace for each space.

use a different BanachSpace for each space.

julia> a = Sequence(Taylor(1)^2, [1.0, 2.0, 3.0, 4.0]);
julia> norm(a, NormedCartesianSpace(ℓ¹(), ℓ∞()))7.0
julia> norm(a, NormedCartesianSpace((ℓ¹(), ℓ²()), ℓ∞()))5.0

RadiiPolynomial.NormedCartesianSpace — Type

NormedCartesianSpace{T<:Union{BanachSpace,Tuple{Vararg{BanachSpace}}},S<:BanachSpace} <: BanachSpace

Cartesian Banach space.

Fields:

inner :: T

outer :: S