Special operators

Projection

julia> using RadiiPolynomial
julia> A = LinearOperator(Taylor(1) ⊗ Chebyshev(1), Taylor(1) ⊗ Chebyshev(1), [1.0 0.0 0.0 0.0 ; 0.0 1.0 0.0 0.0 ; 0.0 0.0 1.0 0.0 ; 0.0 0.0 0.0 1.0])LinearOperator : Taylor(1) ⊗ Chebyshev(1) → Taylor(1) ⊗ Chebyshev(1) with coefficients Matrix{Float64}:
 1.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0
 0.0  0.0  1.0  0.0
 0.0  0.0  0.0  1.0
julia> project(A, Taylor(1) ⊗ Chebyshev(2), Taylor(2) ⊗ Chebyshev(1), Float64)LinearOperator : Taylor(1) ⊗ Chebyshev(2) → Taylor(2) ⊗ Chebyshev(1) with coefficients Matrix{Float64}:
 1.0  0.0  0.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0

Multiplication

RadiiPolynomial.Multiplication — Type

Multiplication{<:Sequence{<:SequenceSpace}}

Multiplication operator associated to a Sequence in a SequenceSpace.

julia> using RadiiPolynomial
julia> A = Multiplication(Sequence(Taylor(1) ⊗ Fourier(1, 1.0), [1.0, 2.0, 3.0, 4.0, 5.0, 6.0]))Multiplication{Sequence{TensorSpace{Tuple{Taylor, Fourier{Float64}}}, Vector{Float64}}}(Sequence{TensorSpace{Tuple{Taylor, Fourier{Float64}}}, Vector{Float64}}(TensorSpace{Tuple{Taylor, Fourier{Float64}}}((Taylor(1), Fourier{Float64}(1, 1.0))), [1.0, 2.0, 3.0, 4.0, 5.0, 6.0]))
julia> project(A, Taylor(1) ⊗ Fourier(1, 1.0), Taylor(2) ⊗ Fourier(0, 1.0), Float64)LinearOperator : Taylor(1) ⊗ Fourier{Float64}(1, 1.0) → Taylor(2) ⊗ Fourier{Float64}(0, 1.0) with coefficients Matrix{Float64}:
 5.0  0.0  3.0  0.0  1.0  0.0
 6.0  5.0  4.0  3.0  2.0  1.0
 0.0  6.0  0.0  4.0  0.0  2.0

Derivation and integration

RadiiPolynomial.Derivative — Type

Derivative{T<:Union{Int,Tuple{Vararg{Int}}}}

Generic derivative operator.

See also: differentiate and differentiate!.

Examples

julia> Derivative(1)
Derivative{Int64}(1)

julia> Derivative(1, 2)
Derivative{Tuple{Int64, Int64}}((1, 2))

RadiiPolynomial.differentiate — Function

differentiate(a::Sequence, α=1)

Computes the α-th derivative of a.

See also: differentiate! and Derivative.

RadiiPolynomial.differentiate! — Function

differentiate!(c::Sequence, a::Sequence, α=1)

Computes the α-th derivative of a. The result is stored in c by overwritting it.

See also: differentiate and Derivative.

julia> using RadiiPolynomial
julia> project(Derivative(1), Taylor(2), Taylor(1), Float64)LinearOperator : Taylor(2) → Taylor(1) with coefficients SparseMatrixCSC{Float64, Int64}:
  ⋅   1.0   ⋅
  ⋅    ⋅   2.0
julia> project(Derivative((1, 1)), Taylor(2) ⊗ Fourier(1, 1.0), Taylor(1) ⊗ Fourier(1, 1.0), Complex{Float64})LinearOperator : Taylor(2) ⊗ Fourier{Float64}(1, 1.0) → Taylor(1) ⊗ Fourier{Float64}(1, 1.0) with coefficients SparseMatrixCSC{ComplexF64, Int64}:
      ⋅       0.0 - 1.0im       ⋅       …       ⋅            ⋅
      ⋅            ⋅       0.0 - 2.0im          ⋅            ⋅
      ⋅            ⋅            ⋅               ⋅            ⋅
      ⋅            ⋅            ⋅               ⋅            ⋅
      ⋅            ⋅            ⋅          0.0 + 1.0im       ⋅
      ⋅            ⋅            ⋅       …       ⋅       0.0 + 2.0im

RadiiPolynomial.Integral — Type

Integral{T<:Union{Int,Tuple{Vararg{Int}}}}

Generic integral operator.

See also: integrate and integrate!.

Examples

julia> Integral(1)
Integral{Int64}(1)

julia> Integral(1, 2)
Integral{Tuple{Int64, Int64}}((1, 2))

RadiiPolynomial.integrate — Function

integrate(a::Sequence, α=1)

Computes the α-th integral of a.

See also: integrate! and Integral.

RadiiPolynomial.integrate! — Function

integrate!(c::Sequence, a::Sequence, α=1)

Computes the α-th integral of a. The result is stored in c by overwritting it.

See also: integrate and Integral.

julia> using RadiiPolynomial
julia> project(Integral(1), Taylor(2), Taylor(3), Float64)LinearOperator : Taylor(2) → Taylor(3) with coefficients SparseMatrixCSC{Float64, Int64}:
  ⋅    ⋅    ⋅
 1.0   ⋅    ⋅
  ⋅   0.5   ⋅
  ⋅    ⋅   0.3333333333333333
julia> project(Integral((1, 1)), Taylor(1) ⊗ Fourier(1, 1.0), Taylor(2) ⊗ Fourier(1, 1.0), Complex{Float64})LinearOperator : Taylor(1) ⊗ Fourier{Float64}(1, 1.0) → Taylor(2) ⊗ Fourier{Float64}(1, 1.0) with coefficients SparseMatrixCSC{ComplexF64, Int64}:
      ⋅            ⋅            ⋅            ⋅            ⋅            ⋅
 0.0 + 1.0im       ⋅            ⋅            ⋅            ⋅            ⋅
      ⋅       0.0 + 0.5im       ⋅            ⋅            ⋅            ⋅
      ⋅            ⋅            ⋅            ⋅            ⋅            ⋅
      ⋅            ⋅       0.0 + 0.0im       ⋅            ⋅            ⋅
      ⋅            ⋅            ⋅       0.0 + 0.0im       ⋅            ⋅
      ⋅            ⋅            ⋅            ⋅            ⋅            ⋅
      ⋅            ⋅            ⋅            ⋅       0.0 - 1.0im       ⋅
      ⋅            ⋅            ⋅            ⋅            ⋅       0.0 - 0.5im

Evaluation

RadiiPolynomial.Evaluation — Type

Evaluation{T<:Union{Nothing,Number,Tuple{Vararg{Union{Nothing,Number}}}}}

Generic evaluation operator. A value of nothing indicates that no evaluation should be performed.

See also: evaluate and evaluate!.

Examples

julia> Evaluation(1.0)
Evaluation{Float64}(1.0)

julia> Evaluation(1.0, nothing, 2.0)
Evaluation{Tuple{Float64, Nothing, Float64}}((1.0, nothing, 2.0))

RadiiPolynomial.evaluate — Function

evaluate(a::Sequence, x)

Evaluates a at x.

See also: evaluate! and Evaluation.

RadiiPolynomial.evaluate! — Function

evaluate!(c::Sequence, a::Sequence, x)

Evaluates a at x. The result is stored in c by overwritting it.

See also: evaluate and Evaluation.

julia> using RadiiPolynomial
julia> project(Evaluation(0.5), Taylor(2), Taylor(0), Float64)LinearOperator : Taylor(2) → Taylor(0) with coefficients Matrix{Float64}:
 1.0  0.5  0.25
julia> project(Evaluation((0.5, nothing)), Taylor(2) ⊗ Fourier(1, 1.0), Taylor(0) ⊗ Fourier(1, 1.0), Float64)LinearOperator : Taylor(2) ⊗ Fourier{Float64}(1, 1.0) → Taylor(0) ⊗ Fourier{Float64}(1, 1.0) with coefficients Matrix{Float64}:
 1.0  0.5  0.25  0.0  0.0  0.0   0.0  0.0  0.0
 0.0  0.0  0.0   1.0  0.5  0.25  0.0  0.0  0.0
 0.0  0.0  0.0   0.0  0.0  0.0   1.0  0.5  0.25

Scale

RadiiPolynomial.Scale — Type

Scale{<:Union{Number,Tuple{Vararg{Number}}}}

Generic scale operator.

See also: scale and scale!.

Examples

julia> Scale(1.0)
Scale{Float64}(1.0)

julia> Scale(1.0, 2.0)
Scale{Tuple{Float64, Float64}}((1.0, 2.0))

RadiiPolynomial.scale — Function

scale(a::Sequence, γ)

Scales a by a factor γ.

See also: scale! and Scale.

RadiiPolynomial.scale! — Function

scale!(c::Sequence, a::Sequence, γ)

Scales a by a factor γ. The result is stored in c by overwritting it.

See also: scale and Scale.

julia> using RadiiPolynomial
julia> project(Scale(2.0), Taylor(2), Taylor(2), Float64)LinearOperator : Taylor(2) → Taylor(2) with coefficients SparseMatrixCSC{Float64, Int64}:
 1.0   ⋅    ⋅
  ⋅   2.0   ⋅
  ⋅    ⋅   4.0

Shift

RadiiPolynomial.Shift — Type

Shift{T<:Union{Number,Tuple{Vararg{Number}}}}

Generic shift operator.

See also: shift and shift!.

Examples

julia> Shift(1.0)
Shift{Float64}(1.0)

julia> Shift(1.0, 2.0)
Shift{Tuple{Float64, Float64}}((1.0, 2.0))

RadiiPolynomial.shift — Function

shift(a::Sequence, τ)

Shifts a by τ.

See also: shift! and Shift.

RadiiPolynomial.shift! — Function

shift!(c::Sequence, a::Sequence, τ)

Shifts a by τ. The result is stored in c by overwritting it.

See also: shift and Shift.

julia> using RadiiPolynomial
julia> project(Shift(π/2), Fourier(1, 1.0), Fourier(1, 1.0), Complex{Float64})LinearOperator : Fourier{Float64}(1, 1.0) → Fourier{Float64}(1, 1.0) with coefficients SparseMatrixCSC{ComplexF64, Int64}:
 6.123233995736766e-17 - 1.0im       ⋅                              ⋅
                        ⋅       1.0 + 0.0im                         ⋅
                        ⋅            ⋅       6.123233995736766e-17 + 1.0im